nass-788: rename hook and move to file
